+++ app/controllers/frontend/characters/items_controller.rb
@@ -78,13 +78,14 @@ module Frontend
         params.require(:character_item).permit!.to_h
       end
 
-      def characters_relation
+      def characters_relation # rubocop: disable Metrics/AbcSize
         case params[:provider]
         when 'dnd5', 'dnd2024' then authorized_scope(Character.all).dnd
         when 'pathfinder2' then authorized_scope(Character.all).pathfinder2
         when 'daggerheart' then authorized_scope(Character.all).daggerheart
         when 'dc20' then authorized_scope(Character.all).dc20
         when 'fallout' then authorized_scope(Character.all).fallout
+        when 'cosmere' then authorized_scope(Character.all).cosmere
         else Character.none
         end
       end
@@ -96,6 +97,7 @@ module Frontend
         when 'daggerheart' then ::Item.daggerheart
         when 'dc20' then ::Item.dc20
         when 'fallout' then ::Item.fallout
+        when 'cosmere' then ::Item.cosmere
         else []
         end
       end