+++ db/seeds_prod.rb
@@ -23,3 +23,10 @@ end
     item ? item.update!(data) : ::Pathfinder2::Item.create!(data)
   end
 end
+
+['weapon.json', 'armor.json', 'items.json'].each do |filename|
+  JSON.parse(File.read("db/data_prod/cosmere/#{filename}")).each do |data|
+    item = ::Cosmere::Item.find_by(slug: data['slug'])
+    item ? item.update!(data) : ::Cosmere::Item.create!(data)
+  end
+end