+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - selecting ancestry and cultures for Cosmere characters
+- rendering equipment for cosmere characters
 
 ## [0.4.29] - 2026-04-21
 ### Added